Created new function for getHighestUnwinnableCardIndexInSuite()
